Combine ReactRootView and ReactRootControl
