+++ vnext/IntegrationTests/TestRootView.cpp
@@ -8,8 +8,6 @@ namespace Microsoft::React::Test {
 TestRootView::TestRootView(const std::string &name, int64_t width, int64_t height)
     : m_name{name}, m_width{width}, m_height{height} {}
 
-void TestRootView::ResetView() {}
-
 std::string TestRootView::JSComponentName() const noexcept {
   return m_name;
 }
