--- vnext/Microsoft.ReactNative/Views/ReactRootControl.h
@@ -1,155 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-#pragma once
-#include <memory>
-
-#include <IReactInstance.h>
-#include <object/unknownObject.h>
-#include "IXamlRootView.h"
-#include "ReactHost/React.h"
-#include "SIPEventHandler.h"
-#include "TouchEventHandler.h"
-#include "Views/KeyboardEventHandler.h"
-
-namespace winrt {
-using namespace Windows::UI;
-using namespace Windows::UI::Core;
-using namespace xaml;
-using namespace xaml::Controls;
-using namespace xaml::Input;
-using namespace Windows::Foundation;
-using namespace xaml::Media;
-} // namespace winrt
-
-namespace Microsoft::ReactNative {
-
-// A class that implements IXamlRootView and IXamlReactControl interfaces.
-// It keeps a weak reference to the XAML ReactRootView.
-// TODO: consider to remove this class in favor of ReactRootView XAML control.
-struct ReactRootControl final : std::enable_shared_from_this<ReactRootControl>, IXamlRootView, IXamlReactControl {
-  ReactRootControl(Microsoft::ReactNative::XamlView const &rootView) noexcept;
-  ~ReactRootControl() noexcept;
-
- public: // IXamlRootView
-  Mso::React::IReactContext *GetReactContext() const noexcept override;
-  Microsoft::ReactNative::XamlView GetXamlView() const noexcept override;
-  void SetJSComponentName(std::string &&mainComponentName) noexcept override;
-  void SetInitialProps(folly::dynamic &&initialProps) noexcept override;
-  void AttachRoot() noexcept override;
-  void DetachRoot() noexcept override;
-  std::shared_ptr<IXamlReactControl> GetXamlReactControl() const noexcept override;
-
- public: // IReactRootView
-  void ResetView() noexcept override;
-  std::string JSComponentName() const noexcept override;
-  int64_t GetActualHeight() const noexcept override;
-  int64_t GetActualWidth() const noexcept override;
-  int64_t GetTag() const noexcept override;
-  void SetTag(int64_t tag) noexcept override;
-
- public: // IXamlReactControl
-  void blur(Microsoft::ReactNative::XamlView const &xamlView) noexcept override;
-
- public:
-  //! property ReactViewHost : Mso::React::IReactViewHost
-  Mso::React::IReactViewHost *ReactViewHost() noexcept;
-  void ReactViewHost(Mso::React::IReactViewHost *viewHost) noexcept;
-
- public: // IReactViewInstance UI-thread implementation
-  void InitRootView(
-      Mso::CntPtr<Mso::React::IReactInstance> &&reactInstance,
-      Mso::React::ReactViewOptions &&viewOptions) noexcept;
-  void UpdateRootView() noexcept;
-  void UninitRootView() noexcept;
-
- private:
-  void PrepareXamlRootView(Microsoft::ReactNative::XamlView const &rootView) noexcept;
-  void EnsureFocusSafeHarbor() noexcept;
-  void UpdateRootViewInternal() noexcept;
-
-  void ShowInstanceLoading() noexcept;
-  void ShowInstanceWaiting() noexcept;
-  void ShowInstanceLoaded() noexcept;
-  void ShowInstanceError() noexcept;
-
-  void AttachBackHandlers(Microsoft::ReactNative::XamlView const &rootView) noexcept;
-  void RemoveBackHandlers() noexcept;
-  bool OnBackRequested() noexcept;
-
- private:
-  void ClearLoadingUI() noexcept;
-  void EnsureLoadingUI() noexcept;
-
-  int64_t m_rootTag{-1};
-
-  std::shared_ptr<Microsoft::ReactNative::TouchEventHandler> m_touchEventHandler;
-  std::shared_ptr<Microsoft::ReactNative::SIPEventHandler> m_SIPEventHandler;
-  std::shared_ptr<Microsoft::ReactNative::PreviewKeyboardEventHandlerOnRoot> m_previewKeyboardEventHandlerOnRoot;
-
-  Mso::DispatchQueue m_uiQueue;
-  Mso::CntPtr<Mso::React::IReactViewHost> m_reactViewHost;
-  Mso::WeakPtr<Mso::React::IReactInstance> m_weakReactInstance;
-  Mso::CntPtr<Mso::React::IReactContext> m_context;
-  std::unique_ptr<Mso::React::ReactOptions> m_reactOptions;
-  std::unique_ptr<Mso::React::ReactViewOptions> m_reactViewOptions;
-
-  bool m_isInitialized{false};
-  bool m_isJSViewAttached{false};
-
-  // Visual tree to support safe harbor
-  // m_rootView
-  //  safe harbor
-  //  m_xamlRootView
-  //    JS created children
-  winrt::weak_ref<Microsoft::ReactNative::XamlView> m_weakRootView{nullptr};
-  winrt::weak_ref<Microsoft::ReactNative::XamlView> m_weakXamlRootView{nullptr};
-
-  xaml::Controls::ContentControl m_focusSafeHarbor{nullptr};
-  xaml::Controls::ContentControl::LosingFocus_revoker m_focusSafeHarborLosingFocusRevoker{};
-  xaml::FrameworkElement::SizeChanged_revoker m_rootSizeChangedRevoker{};
-  winrt::Grid m_greenBoxGrid{nullptr};
-  winrt::TextBlock m_waitingTextBlock{nullptr};
-  winrt::SystemNavigationManager::BackRequested_revoker m_backRequestedRevoker{};
-};
-
-//! This class ensures that we access ReactRootView from UI thread.
-struct ReactViewInstance : public Mso::UnknownObject<Mso::RefCountStrategy::WeakRef, Mso::React::IReactViewInstance> {
-  ReactViewInstance(std::weak_ptr<ReactRootControl> &&weakRootControl, Mso::DispatchQueue const &uiQueue) noexcept;
-
-  Mso::Future<void> InitRootView(
-      Mso::CntPtr<Mso::React::IReactInstance> &&reactInstance,
-      Mso::React::ReactViewOptions &&viewOptions) noexcept override;
-  Mso::Future<void> UpdateRootView() noexcept override;
-  Mso::Future<void> UninitRootView() noexcept override;
-
- private:
-  template <class TAction>
-  Mso::Future<void> PostInUIQueue(TAction &&action) noexcept;
-
- private:
-  std::weak_ptr<ReactRootControl> m_weakRootControl;
-  Mso::DispatchQueue m_uiQueue;
-};
-
-//===========================================================================
-// ReactViewInstance inline implementation
-//===========================================================================
-
-template <class TAction>
-inline Mso::Future<void> ReactViewInstance::PostInUIQueue(TAction &&action) noexcept {
-  // ReactViewInstance has shorter lifetime than ReactRootControl. Thus, we capture this WeakPtr.
-  return Mso::PostFuture(
-      m_uiQueue, [weakThis = Mso::WeakPtr{this}, action{std::forward<TAction>(action)}]() mutable noexcept {
-        if (auto strongThis = weakThis.GetStrongPtr()) {
-          if (auto rootControl = strongThis->m_weakRootControl.lock()) {
-            action(*rootControl);
-            return Mso::Maybe<void>{};
-          }
-        }
-
-        return Mso::CancellationErrorProvider().MakeMaybe<void>();
-      });
-}
-
-} // namespace Microsoft::ReactNative
